[Refactor] Player기능 -> PlayerMovement로 이동, Player 프리팹 수정
